Refactored io readers
