+++ tsdat/io/readers/__init__.py
@@ -1,0 +1,14 @@
+from .csv_reader import CSVReader
+from .netcdf_reader import NetCDFReader
+from .parquet_reader import ParquetReader
+from .tar_reader import TarReader
+from .zarr_reader import ZarrReader
+from .zip_reader import ZipReader
+
+__all__ = [
+    "CSVReader",
+    "NetCDFReader",
+    "ParquetReader",
+    "ZarrReader",
+    "ZipReader",
+]
